+++ tidy3d/plugins/waveguide/__init__.py
@@ -1,0 +1,1 @@
+from .rectangular_dielectric import RectangularDielectric